moved my stuff to its own page
