+++ packages/client/src/app/pages/my-stuff/blog-form/blog-form.component.less
@@ -1,9 +1,3 @@
-div.tools {
-    border: 1px solid var(--site-borders);
-    border-radius: 4px;
-    width: 98.8%;
-}
-
 div.form-container {
     width: 85%;
     margin: 0 auto;
