+++ packages/client/src/app/pages/my-stuff/content-item/content-item.component.ts
