+++ packages/client/src/app/pages/my-stuff/my-stuff.component.less
@@ -1,9 +1,3 @@
-div.tools {
-    border: 1px solid var(--site-borders);
-    border-radius: 4px;
-    width: 98.8%;
-}
-
 div.header {
     width: 100%;
     border-bottom: 1px solid var(--site-borders);
