+++ packages/client/src/app/pages/my-stuff/viewmodels/content-item.model.ts
