+++ packages/client/src/app/pages/my-stuff/views/view-poetry/view-poetry.component.less
@@ -1,9 +1,3 @@
-div.tools {
-    border: 1px solid var(--site-borders);
-    border-radius: 4px;
-    width: 98.8%;
-}
-
 div.view-container {
     width: 95%;
     margin: 0 auto;
