+++ packages/client/src/app/pages/my-stuff/views/view-poetry/view-poetry.component.ts
