+++ packages/client/src/app/pages/my-stuff/views/view-prose/view-prose.component.html
