+++ packages/client/src/app/pages/portfolio/portfolio.component.html
@@ -61,16 +61,7 @@
                         <i-feather class="end-point" name="chevron-right"></i-feather>
                     </a>
                 </li>
-                <li *ngIf="currentUser && currentUserIsSame()">
-                    <a [routerLink]="['my-stuff']" routerLinkActive="router-link-active">
-                        <span class="link-icon">
-                            <i-feather name="hard-drive"></i-feather>
-                        </span>
-                        My Stuff
-                        <i-feather class="end-point" name="chevron-right"></i-feather>
-                    </a>
-                </li>
-                <!--<li>
+                <li>
                     <a [routerLink]="['works']" routerLinkActive="router-link-active">
                         <span class="link-icon">
                             <i-feather name="feather"></i-feather>
@@ -97,7 +88,7 @@
                         </ng-template>
                         <i-feather class="end-point" name="chevron-right"></i-feather>
                     </a>
-                </li>-->
+                </li>
                 <li>
                     <a [routerLink]="['collections']" routerLinkActive="router-link-active">
                         <span class="link-icon">
